back-end/requirements.txt
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
fastapi==0.70.1
fastapi-jwt-auth==0.5.0
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.1
pycparser==2.21
pydantic==1.8.2
PyJWT==1.7.1
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
